+++ local-notifications/rollup.config.js
@@ -1,0 +1,22 @@
+export default {
+  input: 'dist/esm/index.js',
+  output: [
+    {
+      file: 'dist/plugin.js',
+      format: 'iife',
+      name: 'capacitorHaptics',
+      globals: {
+        '@capacitor/core': 'capacitorLocalNotifications',
+      },
+      sourcemap: true,
+      inlineDynamicImports: true,
+    },
+    {
+      file: 'dist/plugin.cjs.js',
+      format: 'cjs',
+      sourcemap: true,
+      inlineDynamicImports: true,
+    },
+  ],
+  external: ['@capacitor/core'],
+};